test: move BrowserView setBackgroundColor screen-capture tests to spec/serial
These capture the primary display and assert center-pixel colour, which
breaks when a parallel worker has a window visible.
